requirements.txt
aerich==0.7.2
aiofiles==0.7.0
aiomysql==0.2.0
aioredis==2.0.1
aiosqlite==0.16.1
anyio==3.7.1
APScheduler==3.10.4
asgiref==3.3.4
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
captcha==0.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.6
cryptography==3.4.7
dictdiffer==0.9.0
exceptiongroup==1.1.3
extensions==0.4
fastapi==0.103.0
fastapi-login==1.6.3
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.0
h11==0.12.0
idna==2.10
iso8601==1.1.0
itsdangerous==1.1.0
Jinja2==3.0.1
loguru==0.5.3
MarkupSafe==2.0.1
marshmallow==3.12.1
outcome==1.2.0
passlib==1.7.4
Pillow==8.2.0
psutil==5.8.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.8.0
PyMySQL==1.0.2
PyPika==0.48.9
pypika-tortoise==0.1.6
python-multipart==0.0.5
pytz==2020.5
redis==5.0.0
redlock==1.2.0
requests==2.25.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.18
sqlparse==0.4.1
starlette==0.27.0
tomlkit==0.12.1
tortoise-orm==0.20.0
trio==0.21.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.5
uvicorn==0.14.0
Werkzeug==2.0.1
win32-setctime==1.1.0
